feat: browser-test
