--- .eslintrc.json
@@ -1,26 +1,0 @@
-{
-  "env": {
-      "browser": true,
-      "es2021": true,
-      "node": true
-  },
-  "extends": [
-      "airbnb-base"
-  ],
-  "parser": "@typescript-eslint/parser",
-  "parserOptions": {
-      "ecmaVersion": 12,
-      "sourceType": "module"
-  },
-  "plugins": [
-      "@typescript-eslint"
-  ],
-  "rules": {
-      "import/prefer-default-export": "off",
-      "import/no-unresolved": "off",
-      "no-restricted-syntax": "off",
-      "import/extensions": "off",
-      "no-undef": "off",
-      "no-await-in-loop": "off"
-  }
-}
