+++ package.json
@@ -16,7 +16,8 @@
     "test": "jest",
     "lint": "eslint lib/* __tests__/* --ext .ts",
     "lint:fix": "eslint lib/* __tests__/* --ext .ts --fix",
-    "semantic-release": "semantic-release"
+    "semantic-release": "semantic-release",
+    "build": "webpack --config webpack.config.js"
   },
   "repository": {
     "type": "git",
@@ -31,16 +32,14 @@
   "homepage": "https://github.com/jeswr/demo-perf-tests/README.md",
   "devDependencies": {
     "@types/jest": "^29.5.12",
-    "@typescript-eslint/eslint-plugin": "^8.3.0",
-    "@typescript-eslint/parser": "^8.3.0",
-    "eslint": "^9.9.1",
-    "eslint-config-airbnb-base": "^15.0.0",
-    "eslint-plugin-import": "^2.29.1",
+    "@types/selenium-webdriver": "^4.1.26",
     "jest": "^29.7.0",
     "pre-commit": "^1.2.2",
     "semantic-release": "^24.1.0",
     "ts-jest": "^29.2.5",
-    "typescript": "^5.5.4"
+    "typescript": "^5.5.4",
+    "webpack": "^5.94.0",
+    "webpack-cli": "^5.1.4"
   },
   "jest": {
     "preset": "ts-jest",
@@ -84,12 +83,10 @@
   },
   "dependencies": {
     "@comunica/reasoning-context-entries": "^0.2.0",
-    "deep-taxonomy-benchmark": "^1.2.0",
+    "deep-taxonomy-benchmark": "^2.0.0",
     "eyereasoner": "^16.20.0",
     "hylar-core": "^1.2.2",
-    "n3": "github:rdfjs/N3.js#a13e50e20c18eb0bc869db2a7c09fe552c653a4e"
-  },
-  "overrides": {
-    "n3": "$n3"
+    "n3": "^1.21.0",
+    "selenium-webdriver": "^4.24.0"
   }
 }
