+++ webpage.html
@@ -1,0 +1,16 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>My Library</title>
+</head>
+<body>
+  <h1>My Library</h1>
+  <script src="bundle/index.js"></script>
+  <script>
+    globalThis.process = { env: { N3: true } }
+    perf.runAll();
+  </script>
+</body>
+</html>
